public/server/requirements.txt
setuptools==41.0.0
numpy==1.19.1
matplotlib==3.3.0
Keras==2.4.3
opencv-python==4.3.0.36
opencv-contrib-python==4.3.0.36
scikit-learn==0.23.1
scikit-image==0.17.2
scipy==1.4.1
mysql-connector-python==8.0.21
Pillow==7.2.0
pickle-mixin==1.0.2
glob3==0.0.1
Flask==1.1.2
Flask-Cors==3.0.8
mpld3==0.5.1
moviepy==1.0.3
tensorflow==2.2.0
jsonify==0.5.0
